pyproject.toml
[tool.poetry]
name = "sns-project"
version = "0.1.0"
description = ""
authors = ["Kwon Sang <kwonsg514@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "4.2"
django-bootstrap-v5 = "^1.0.11"
django-crispy-forms = "^2.1"
crispy-bootstrap4 = "^2023.1"
pillow = "^10.1.0"
django-taggit = "^5.0.1"


[tool.poetry.group.dev.dependencies]
django-debug-toolbar = "^4.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
